+++ backend/src/core/arrays.py
