+++ backend/src/core/backend_config.py
@@ -1,7 +1,7 @@
 import ctypes
 import os
 
-import src.arrays as arrays_manager
+import src.core.arrays as arrays_manager
 import src.sorts.helpers as helpers
 import src.sorts.python.python_sorts as sort_impls
 
@@ -12,7 +12,7 @@ pySorts backend API address
 
 # load the c sorting algorithm implementations and expose in path
 script_dir = os.path.abspath(os.path.dirname(__file__))
-lib_path = os.path.join(script_dir, "sorts/c/cSorts.so")
+lib_path = os.path.join(script_dir, "../sorts/c/cSorts.so")
 cSorts = ctypes.cdll.LoadLibrary(lib_path)
 
 """
